Added convertions methods in ProductServiceImpl

+++ src/main/java/backend2/fakestoreapp/DTO/RatingDTO.java
@@ -7,6 +7,6 @@ import java.math.BigDecimal;
 @Getter
 @Setter
 public class RatingDTO {
-    private BigDecimal rate;
+    private double rate;
     private Integer count;
 }
